Move app to root and format
