+++ README.md
@@ -1,1 +1,1 @@
-# Mortein Frontend Application
+# Plant App - Front End
