+++ index.html
@@ -3,9 +3,10 @@
     <head>
         <meta charset="UTF-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title>Mortein Frontend Application</title>
+        <title>Plant App</title>
     </head>
     <body>
-        <p>Hello, World!</p>
+        <div id="root"></div>
+        <script type="module" src="/src/main.jsx"></script>
     </body>
 </html>
