--- plant-app/README.md
@@ -1,1 +1,0 @@
-# Plant App - Front End
