--- plant-app/src/App.css
@@ -1,96 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-    color: #7ea07f;
-    font-family: Verdana, Geneva, Tahoma, sans-serif;
-}
-
-.detail {
-    background-color: #fdfff2;
-}
-
-.navbar {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    background-color: #eff4de;
-    padding: 20px;
-    font-size: 0.8rem;
-}
-
-.plant-main {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: 20px;
-    font-size: 1rem;
-}
-
-.plant-image {
-    width: 100px;
-    height: 100px;
-    border-radius: 50%;
-    border: 1px solid black;
-    margin: 20px;
-}
-
-.plant-health {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.plant-health-title {
-    padding: 20px;
-    margin-bottom: 10px;
-    font-size: 1.2rem;
-}
-
-.meter {
-    display: flex;
-    align-items: center;
-    width: 80vw;
-    height: 30px;
-    background-color: white;
-    border-radius: 12px;
-    position: relative;
-    border: 1px solid black;
-}
-
-.level {
-    margin-top: 0;
-    height: 100%;
-    border-radius: 12px;
-    position: absolute;
-}
-
-.sunlight .level {
-    background-color: #ededa7;
-}
-
-.moisture .level {
-    background-color: #b0cce8;
-}
-
-.bar-heading,
-.vibration-heading {
-    font-size: 1rem;
-
-}
-
-.bar-heading {
-    margin-top: 30px;
-}
-
-.vibration {
-    display: flex;
-    margin: 20px;
-    padding: 20px;
-    gap: 20vw;
-}
-
-button {
-    width: 50px;
-    font-size: 1rem;
-}
