--- plant-app/src/pages/Guide.jsx
@@ -1,10 +1,0 @@
-function Guide() {
-    
-    return (
-        <div className="guide-page">
-            <h1>Guide Page</h1>
-        </div>
-    )
-}
-
-export default Guide
