+++ src/components/PlantIcon.jsx
@@ -1,0 +1,11 @@
+import { Link } from "react-router-dom";
+
+function PlantIcon(props) {
+    return (
+        <div className="plant-icon">
+            <h1>{props.name}</h1>
+        </div>
+    );
+}
+
+export default PlantIcon;
