+++ src/pages/Guide.jsx
@@ -1,0 +1,9 @@
+function Guide() {
+    return (
+        <div className="guide-page">
+            <h1>Guide Page</h1>
+        </div>
+    );
+}
+
+export default Guide;
