+++ src/pages/Login.jsx
@@ -1,0 +1,9 @@
+function Login() {
+    return (
+        <div className="login-page">
+            <h1>Login Page</h1>
+        </div>
+    );
+}
+
+export default Login;
